added eager Loading with include and get current instructor
Changes the method name to EditPost because the signature is now the
same as the HttpGet method (the ActionName attribute specifies that the
/Edit/ URL is still used).
 Gets the current Instructor entity from the database using eager
loading for the OfficeAssignment navigation property. This is the same
as what you did in the HttpGet Edit method.
 Updates the retrieved Instructor entity with values from the model
binder. The TryUpdateModel overload used enables you to whitelist the
properties you want to include. This prevents over-posting,